# raft: optimize ComputeLeadSupportUntil