--- a/pkg/raft/quorum/majority.go
+++ b/pkg/raft/quorum/majority.go
@@ -193,10 +193,10 @@ func (c MajorityConfig) VoteResult(votes map[pb.PeerID]bool) VoteResult {
 	return VoteLost
 }
 
-// LeadSupportExpiration takes a mapping of timestamps peers have promised a
+// LeadSupportExpiration takes a slice of timestamps peers have promised a
 // fortified leader support until and returns the timestamp until which the
 // leader is guaranteed support until.
-func (c MajorityConfig) LeadSupportExpiration(supported map[pb.PeerID]hlc.Timestamp) hlc.Timestamp {
+func (c MajorityConfig) LeadSupportExpiration(support []hlc.Timestamp) hlc.Timestamp {
 	if len(c) == 0 {
 		// There are no peers in the config, and therefore no leader, so we return
 		// MaxTimestamp as a sentinel value. This also plays well with joint quorums
@@ -206,33 +206,7 @@ func (c MajorityConfig) LeadSupportExpiration(supported map[pb.PeerID]hlc.Timest
 	}
 
 	n := len(c)
-
-	// Use an on-stack slice whenever n <= 7 (otherwise we alloc). The assumption
-	// is that running with a replication factor of >7 is rare, and in cases in
-	// which it happens, performance is less of a concern (it's not like
-	// performance implications of an allocation here are drastic).
-	var stk [7]hlc.Timestamp
-	var srt []hlc.Timestamp
-	if len(stk) >= n {
-		srt = stk[:n]
-	} else {
-		srt = make([]hlc.Timestamp, n)
-	}
-
-	{
-		// Fill the slice with Timestamps for peers in the configuration. Any unused
-		// slots will be left as empty Timestamps  for our calculation. We fill from
-		// the right (since the zeros will end up on the left after sorting anyway).
-		i := n - 1
-		for id := range c {
-			if e, ok := supported[id]; ok {
-				srt[i] = e
-				i--
-			}
-		}
-	}
-
-	slices.SortFunc(srt, func(a hlc.Timestamp, b hlc.Timestamp) int {
+	slices.SortFunc(support, func(a hlc.Timestamp, b hlc.Timestamp) int {
 		return a.Compare(b)
 	})
 
@@ -240,9 +214,9 @@ func (c MajorityConfig) LeadSupportExpiration(supported map[pb.PeerID]hlc.Timest
 	// assumption is that if a timestamp is supported by a peer, so are all
 	// timestamps less than that timestamp. For this, we can simply consider the
 	// quorum formed by picking the highest value elements and pick the minimum
-	// from this. In other words, from our sorted (in increasing order) array srt,
-	// we want to move n/2 + 1 to the left from the end (accounting for
+	// from this. In other words, from our sorted (in increasing order) array
+	// support, we want to move n/2 + 1 to the left from the end (accounting for
 	// zero-indexing).
 	pos := n - (n/2 + 1)
-	return srt[pos]
+	return support[pos]
 }