This commit optimizes ComputeLeadSupportUntil by swapping the map that was previously used with a slice (on stack if possible).


Benchmark results before:
```
benchdiff "--run=^BenchmarkComputeLeadSupportUntil$" "--old=HEAD^" "--new=HEAD" "./pkg/raft/tracker"                                                                                                                                                                                                 
                                                                                                                                                                                                                                                                                                     
name                                    old time/op    new time/op    delta                                                                                                                                                                                                                          
ComputeLeadSupportUntil/members=100-24    14.2µs ± 1%     5.0µs ± 2%  -65.17%  (p=0.000 n=9+10)                                                                                                                                                                                                      
ComputeLeadSupportUntil/members=7-24       542ns ± 0%     246ns ± 0%  -54.51%  (p=0.000 n=8+9)                                                                                                                                                                                                       
ComputeLeadSupportUntil/members=5-24       420ns ± 0%     196ns ± 0%  -53.33%  (p=0.000 n=9+9)                                                                                                                                                                                                       
ComputeLeadSupportUntil/members=3-24       300ns ± 0%     144ns ± 0%  -52.14%  (p=0.000 n=8+8)                                                                                                                                                                                                       
ComputeLeadSupportUntil/members=1-24       177ns ± 0%      85ns ± 0%  -51.89%  (p=0.000 n=10+9)                                                                                                                                                                                                      
                                                                                                                                                                                                                                                                                                     
name                                    old alloc/op   new alloc/op   delta                                                                                                                                                                                                                          
ComputeLeadSupportUntil/members=100-24    1.92kB ± 0%    1.79kB ± 0%   -6.72%  (p=0.002 n=8+10)                                                                                                                                                                                                      
ComputeLeadSupportUntil/members=1-24       0.00B          0.00B          ~     (all equal)                                                                                                                                                                                                           
ComputeLeadSupportUntil/members=3-24       0.00B          0.00B          ~     (all equal)                                                                                                                                                                                                           
ComputeLeadSupportUntil/members=5-24       0.00B          0.00B          ~     (all equal)                                                                                                                                                                                                           
ComputeLeadSupportUntil/members=7-24       0.00B          0.00B          ~     (all equal)                                                                                                                                                                                                           
                                                                                                                                                                                                                                                                                                     
name                                    old allocs/op  new allocs/op  delta                                                                                                                                                                                                                          
ComputeLeadSupportUntil/members=100-24      4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.000 n=10+10)                                                                                                                                                                                                     
ComputeLeadSupportUntil/members=1-24        0.00           0.00          ~     (all equal)                                                                                                                                                                                                           
ComputeLeadSupportUntil/members=3-24        0.00           0.00          ~     (all equal)                                                                                                                                                                                                           
ComputeLeadSupportUntil/members=5-24        0.00           0.00          ~     (all equal)                                                                                                                                                                                                           
ComputeLeadSupportUntil/members=7-24        0.00           0.00          ~     (all equal) 
```

Release note: None

Fixes: #141511